changed to XeLaTex, using Adobe's MinionPro and MyriadPro fonts.
If you want to keep using the old fonts or PdfLatex, check the "Language stuff" section and comment/uncomment the parts that you need / do not need. If you want to use the new version, you need to install THIS https://github.com/sebschub/FontPro. The OTF fonts can be found in your Adobe reader installation.